feat(nvim): modular editor utils, clipboard fallback, update docs
Summary:
- Extracted editor behavior into util modules: util/editor.lua, util/autoreload.lua, util/gutter.lua
- Moved terminal/GUI setup to util/terminal.lua; wired into init.lua
- Enabled vim.loader early for built-in Lua module cache
- Switched clipboard to "unnamed,unnamedplus" for cross-platform fallback
- Removed <leader>y clipboard mapping (redundant with unnamedplus)
- Updated README and KEYBINDINGS; added :checkhealth clipboard guidance

Testing:
- Required new modules from init.lua without errors
- Verified &clipboard shows unnamed,unnamedplus
- Searched for stray <leader>y references; none remain

Refs:
- N/A
